services/MediaMixer/requirements.txt
flask
werkzeug
flask-cors
opencv-python
numpy
mss
pillow
click
itsdangerous
jinja2
markupsafe
websockets